refactor(deliberation): simplify component structure and context management
Refactored the deliberation components to use a centralized context for state management. Removed redundant props and streamlined the code by introducing `ClientProviders` and `useDeliberationSpaceContext`. Added type definitions for deliberation-related entities.
