+++ src/app/spaces/[id]/deliberation/_components/thread.tsx
@@ -4,38 +4,24 @@ import React from 'react';
 import SpaceHeader from '../../_components/space_header';
 import SpaceContents from '../../_components/space_contents';
 import SpaceFiles from './space_files';
-import { Thread } from '..';
 import { FileInfo } from '@/lib/api/models/feeds';
-import { SpaceStatus } from '@/lib/api/models/spaces';
+import { useDeliberationSpaceContext } from '../provider.client';
 
-export default function ThreadPage({
-  title,
-  status,
-  thread,
-  setTitle,
-  setThread,
+export default function ThreadPage() {
+  const {
+    isEdit,
+    title,
+    setTitle,
+    thread,
+    setThread,
+    handleGoBack,
+    userType,
+    proposerImage,
+    proposerName,
+    createdAt,
+    status,
+  } = useDeliberationSpaceContext();
 
-  userType,
-  proposerImage,
-  proposerName,
-  createdAt,
-  isEdit,
-
-  onback,
-}: {
-  title: string;
-  status: SpaceStatus;
-  thread: Thread;
-  setThread: (thread: Thread) => void;
-  setTitle: (title: string) => void;
-  userType: number;
-  proposerImage: string;
-  proposerName: string;
-  createdAt: number;
-  isEdit: boolean;
-
-  onback: () => void;
-}) {
   return (
     <div className="flex flex-row w-full gap-5">
       <div className="flex flex-col w-full">
@@ -47,10 +33,8 @@ export default function ThreadPage({
           proposerImage={proposerImage}
           proposerName={proposerName}
           createdAt={createdAt}
-          onback={onback}
-          setTitle={(title: string) => {
-            setTitle(title);
-          }}
+          onback={handleGoBack}
+          setTitle={setTitle}
         />
         <div className="flex flex-col w-full mt-7.5 gap-2.5">
           <SpaceContents
